v1

+++ src/app/zapier/unschedule/deleteCron.ts
@@ -20,6 +20,7 @@ export default async function deleteCron(hookUrl: string): Promise<boolean> {
     .select("*")
     .eq("hook_url", hookUrl);
   if (error || data.length === 0) return false;
+
   // delete cron job
   const headers = {
     "Content-Type": "application/json",
@@ -28,7 +29,22 @@ export default async function deleteCron(hookUrl: string): Promise<boolean> {
   await axios.delete(`https://api.cron-job.org/jobs/${data[0].job_id}`, {
     headers,
   });
+
+  // get account usage
+  const account = await supabase
+    .from("account")
+    .select("*")
+    .eq("email", data[0].email);
+  if (account.error || account.data.length === 0) return false;
+
+  // decreate usage by 1
+  await supabase
+    .from("account")
+    .update({ usage: account.data[0].usage - 1 })
+    .eq("email", data[0].email);
+
   // delete cron from table
   await supabase.from("cron").delete().eq("job_id", data[0].job_id);
+
   return true;
 }
